power: avoid implicit casts of int64_t to double
powerd/policy: As suggested by clang-tidy, avoid implicit of int64_t to
double. Mostly, the best, simple change is to make the casts explicit.

While arguably stylistically slightly better, this change is effectively
a no-op. However, a more useful change would require significant design
work.

BUG=b:258053283
TEST=Local Tests + CQ

Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/4015945
Tested-by: Alan Green <avg@chromium.org>
Commit-Queue: Alan Green <avg@chromium.org>

+++ power_manager/powerd/policy/adaptive_charging_controller.cc
@@ -261,7 +261,8 @@ base::TimeDelta ChargeHistory::GetChargeDurationPercentile(double percentile) {
 
   size_t idx = std::min(
       charge_events_.size() - 1,
-      static_cast<size_t>(std::ceil(percentile * charge_events_.size())));
+      static_cast<size_t>(
+          std::ceil(percentile * static_cast<double>(charge_events_.size()))));
   std::nth_element(durations.begin(), durations.begin() + idx, durations.end());
   return durations[idx];
 }
@@ -1095,7 +1096,7 @@ void AdaptiveChargingController::OnPowerStatusUpdate() {
   if (AtHoldPercent(status.display_battery_percentage)) {
     if (state_ == AdaptiveChargingState::ACTIVE && is_sustain_set_) {
       power_supply_->SetAdaptiveCharging(target_full_charge_time_,
-                                         display_percent_);
+                                         static_cast<double>(display_percent_));
     }
 
     // Since we report metrics on how well the ML model does even if Adaptive
@@ -1165,7 +1166,8 @@ bool AdaptiveChargingController::StartAdaptiveCharging(
   }
 
   started_ = true;
-  report_charge_time_ = status.display_battery_percentage <= hold_percent_;
+  report_charge_time_ =
+      status.display_battery_percentage <= static_cast<double>(hold_percent_);
   base::TimeDelta hold_time_on_ac = charge_history_.GetHoldTimeOnAC();
   base::TimeDelta time_full_on_ac = charge_history_.GetTimeFullOnAC();
   base::TimeDelta time_on_ac = charge_history_.GetTimeOnAC();
@@ -1279,7 +1281,8 @@ bool AdaptiveChargingController::AtHoldPercent(double display_battery_percent) {
   // battery percentage drops below `hold_percent_` - `hold_delta_percent_`.
   // This means that the battery charge can momentarily drop below the lower
   // end of the range we specified.
-  return display_battery_percent >= hold_percent_ - hold_delta_percent_ - 1;
+  return display_battery_percent >=
+         static_cast<double>(hold_percent_ - hold_delta_percent_ - 1);
 }
 
 void AdaptiveChargingController::StartRecheckAlarm(base::TimeDelta delay) {
